split off mtc benchmark and create a plugin system for benchmarks
If anyone does not have MTC in their workspace, they can just
ignore the package and use everything else.

Now any package depending on moveit_benchmark_suite
can define a set of benchmarks to run in terms of
a folder with launch files.
