--- benchmark_suite/launch/mtc_pickplace.launch
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-
-  <param name="mtc_scene" command="$(find xacro)/xacro '$(find moveit_benchmark_suite_resources)/db/openshelf/panda/scene_mesh_hq.urdf.xacro'"/>
-
-  
-  <!-- Run MTC pick and place -->
-  <node name="mtc_benchmark" pkg="moveit_benchmark_suite" type="mtc_runtask" output="screen">
-    <param name="execute" value="true" />
-    <rosparam command="load" file="$(find moveit_benchmark_suite)/config/mtc_pickplace.yaml" />
-  </node>
-</launch>
