+++ mtc/benchmarks/mtc_pickplace.launch
@@ -1,0 +1,21 @@
+<?xml version="1.0"?>
+<launch>
+  <!-- TODO: not implemented in mtc_runtask yet
+  <arg name="output_file" />
+  -->
+
+  <include file="$(find moveit_resources_panda_moveit_config)/launch/demo.launch">
+    <arg name="use_rviz" value="false"/>
+  </include>
+  <!-- add MTC's execution capability as "execute" below might be true -->
+  <param name="move_group/capabilities" value="move_group/ExecuteTaskSolutionCapability" />
+
+  <param name="mtc_scene" command="xacro '$(find moveit_benchmark_suite_resources)/db/openshelf/panda/scene_mesh_hq.urdf.xacro'"/>
+
+  <!-- Run MTC pick and place -->
+  <node name="mtc_benchmark" pkg="moveit_benchmark_suite_mtc" type="mtc_runtask" output="screen" required="true">
+    <param name="execute" value="false" />
+    <param name="output_file" value="$(arg output_file)" />
+    <rosparam command="load" file="$(dirname)/../config/mtc_pickplace.yaml" />
+  </node>
+</launch>
