+++ mtc/config/mtc_pickplace.yaml
@@ -1,5 +1,5 @@
 # Total planning attempts
-max_solutions: 10
+max_solutions: 1
 
 # Planning group and link names
 arm_group_name: "panda_arm"
