+++ mtc/src/mtc_pickplace.cpp
@@ -34,7 +34,8 @@
    Desc:   A demo to show MoveIt Task Constructor in action
 */
 
-#include <moveit_benchmark_suite/test/mtc_pickplace.h>
+#include "mtc_pickplace.h"
+
 #include <rosparam_shortcuts/rosparam_shortcuts.h>
 
 namespace moveit
